Split duplicated reference configs into several spring xml
