+++ dubbo-samples-multi-registry/pom.xml
@@ -29,7 +29,8 @@
     <properties>
         <source.level>1.8</source.level>
         <target.level>1.8</target.level>
-        <dubbo.version>2.7.7</dubbo.version>
+        <dubbo.version>2.7.11</dubbo.version>
+<!--        <dubbo.version>3.0.1-SNAPSHOT</dubbo.version>-->
         <spring.version>4.3.16.RELEASE</spring.version>
         <junit.version>4.12</junit.version>
         <maven-compiler-plugin.version>3.7.0</maven-compiler-plugin.version>
