+++ dubbo-samples-multi-registry/src/main/java/org/apache/dubbo/samples/multi/registry/MultiRegistryConsumer.java
@@ -28,15 +28,26 @@ import org.springframework.context.support.ClassPathXmlApplicationContext;
 public class MultiRegistryConsumer {
 
     public static void main(String[] args) throws Exception {
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("spring/multi-registry-consumer.xml");
+//        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("spring/multi-registry-consumer.xml");
+//        context.start();
+//
+//        DemoService demoServiceFormDefault = (DemoService) context.getBean("demoServiceFormDefault");
+//        HelloService helloServiceFormShanghai = (HelloService) context.getBean("helloServiceFormShanghai");
+//        HelloService helloServiceFormBeijing = (HelloService) context.getBean("helloServiceFormBeijing");
+//
+//        System.out.println(demoServiceFormDefault.get("service form default registry"));
+//        System.out.println(helloServiceFormShanghai.sayHello("service form shanghai registry"));
+//        System.out.println(helloServiceFormBeijing.sayHello("service form beijing registry"));
+
+
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("spring/multi-registry-consumer1.xml");
         context.start();
 
-        DemoService demoServiceFormDefault = (DemoService) context.getBean("demoServiceFormDefault");
-        HelloService helloServiceFormShanghai = (HelloService) context.getBean("helloServiceFormShanghai");
-        HelloService helloServiceFormBeijing = (HelloService) context.getBean("helloServiceFormBeijing");
-
-        System.out.println(demoServiceFormDefault.get("service form default registry"));
-        System.out.println(helloServiceFormShanghai.sayHello("service form shanghai registry"));
-        System.out.println(helloServiceFormBeijing.sayHello("service form beijing registry"));
+        DemoService demoService = context.getBean(DemoService.class);
+        HelloService helloService = context.getBean(HelloService.class);
+        System.out.println(demoService.get("service form default registry"));
+        System.out.println(helloService.sayHello("service form shanghai registry"));
     }
+
+
 }
